Add support for old and new analysis by creating two binary projects
