+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common/build.properties
@@ -1,0 +1,5 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.xml
