+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common/plugin.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+
+</plugin>
