+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common/src/org/palladiosimulator/dataflow/confidentiality/scalability/AnalysisUtils.java
@@ -1,0 +1,11 @@
+package org.palladiosimulator.dataflow.confidentiality.scalability;
+
+public class AnalysisUtils {
+
+    private AnalysisUtils() {
+        // Utility class
+    }
+
+    public static String TEST_MODEL_PROJECT_NAME = "org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels";
+
+}
