+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common/src/org/palladiosimulator/dataflow/confidentiality/scalability/TestRunner.java
@@ -1,0 +1,67 @@
+package org.palladiosimulator.dataflow.confidentiality.scalability;
+
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.ObjectOutputStream;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.apache.log4j.Logger;
+import org.palladiosimulator.dataflow.confidentiality.scalability.result.ScalibilityParameter;
+import org.palladiosimulator.dataflow.confidentiality.scalability.result.ScalibilityTest;
+
+public class TestRunner {
+	private static final int RUNS_PER_STAGE = 10;
+	
+	private final Logger logger = Logger.getLogger(TestRunner.class);
+	
+	private List<ScalibilityTest> tests;
+	private List<ScalibilityParameter> results;
+	private AnalysisExecutor analysisExecutor;
+	private boolean legacy;
+	
+	public TestRunner(List<ScalibilityTest> tests, AnalysisExecutor analysisExecutor, boolean legacy) {
+		this.tests = tests;
+		this.results = new ArrayList<>();
+		this.analysisExecutor = analysisExecutor;
+		this.legacy = legacy;
+	}
+	
+	public void runTests() {
+		this.runTests(0);
+	}
+	
+	public void runTests(int start) {
+		for(int i = start; true; i++) {
+			for(ScalibilityTest test : this.tests) {
+				logger.info("Running test with name " + test.getTestName());
+				this.runTest(test, i);
+			}
+		}
+	}
+	
+	private void runTest(ScalibilityTest test, int index) {
+		for (int j = 0; j < RUNS_PER_STAGE; j++) {
+			int modelSize = test.getModelSize(index);
+			logger.info("Running test with model size " + modelSize);
+			ScalibilityParameter parameter = new ScalibilityParameter(modelSize, test.getTestName(), legacy);
+			test.run(parameter, analysisExecutor);
+			this.results.add(parameter);
+			saveResults(analysisExecutor.getPrefix() + parameter.getTestName());
+		}
+	}
+	
+	private void saveResults(String testName) {
+		try {
+			FileOutputStream fileOutputStream
+		      = new FileOutputStream(testName + ".ser");
+		    ObjectOutputStream objectOutputStream 
+		      = new ObjectOutputStream(fileOutputStream);
+		    objectOutputStream.writeObject(this.results);
+		    objectOutputStream.flush();
+		    objectOutputStream.close();
+		} catch(IOException e) {
+			e.printStackTrace();
+		}
+	}
+}
