+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common/src/org/palladiosimulator/dataflow/confidentiality/scalability/factory/builder/AssemblyAllocationBuilder.java
