+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.new/META-INF/MANIFEST.MF
@@ -1,0 +1,10 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Scalibility Test with New Analysis
+Bundle-SymbolicName: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.new
+Bundle-Version: 1.0.0.qualifier
+Require-Bundle: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common;bundle-version="1.0.0",
+ org.palladiosimulator.dataflow.confidentiality.analysis,
+ org.palladiosimulator.dataflow.confidentiality.analysis.testmodels
+Automatic-Module-Name: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.new
+Bundle-RequiredExecutionEnvironment: JavaSE-17
