+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.old/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Scalibility Test with Old Analysis
+Bundle-SymbolicName: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.old
+Bundle-Version: 1.0.0.qualifier
+Require-Bundle: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.common;bundle-version="1.0.0",
+ org.palladiosimulator.dataflow.confidentiality.transformation.workflow;bundle-version="5.1.0",
+ org.palladiosimulator.dataflow.confidentiality.pcm.workflow;bundle-version="5.1.0",
+ org.palladiosimulator.pcm;bundle-version="5.0.0",
+ org.apache.commons.lang3;bundle-version="3.1.0",
+ org.palladiosimulator.supporting.prolog;bundle-version="5.1.0"
+Automatic-Module-Name: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.old
+Bundle-RequiredExecutionEnvironment: JavaSE-17
