+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.old/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
