+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels/META-INF/MANIFEST.MF
@@ -1,0 +1,10 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Confidentiality Analysis Scaliblity Test Models
+Bundle-SymbolicName: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels
+Bundle-Version: 1.0.0.qualifier
+Automatic-Module-Name: org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.20.0"
+Export-Package: org.palladiosimulator.dataflow.confidentiality.analysis.testmodels
+
