+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels/build.properties
@@ -1,0 +1,7 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               src/,\
+               models/
+
