+++ scalability/org.palladiosimulator.dataflow.confidentiality.analysis.scalibility.testmodels/src/org/palladiosimulator/dataflow/confidentiality/analysis/testmodels/Activator.java
@@ -1,0 +1,8 @@
+
+package org.palladiosimulator.dataflow.confidentiality.analysis.testmodels;
+
+import org.eclipse.core.runtime.Plugin;
+
+public class Activator extends Plugin {
+
+}
